+++ crates/bevy_render/src/primitives/mod.rs
@@ -1,6 +1,8 @@
 use std::borrow::Borrow;
 
-use bevy_ecs::{component::Component, entity::EntityHashMap, prelude::Entity, reflect::ReflectComponent};
+use bevy_ecs::{
+    component::Component, entity::EntityHashMap, reflect::ReflectComponent,
+};
 use bevy_math::{Affine3A, Mat3A, Mat4, Vec3, Vec3A, Vec4, Vec4Swizzles};
 use bevy_reflect::Reflect;
 
@@ -322,7 +324,7 @@ impl CubemapFrusta {
 #[reflect(Component)]
 pub struct CascadesFrusta {
     #[reflect(ignore)]
-    pub frusta: EntityHashMap<Entity, Vec<Frustum>>,
+    pub frusta: EntityHashMap<Vec<Frustum>>,
 }
 
 #[cfg(test)]